refactor: move get_metadata fct to helpers script
